JBIDE-29047: Extract the generic functionality from the experimental runtime plugin   - Move interface 'org.jboss.tools.hibernate.orm.runtime.exp.internal.IDatabaseReader' from 'org.jboss.tools.hibernate.orm.runtime.exp' to 'org.jboss.tools.hibernate.orm.runtime.common.IDatabaseReader' in 'org.jboss.tools.hibernate.orm.runtime.common'
